Re-organize modules based on image type
