+++ src/sknnr_spatial/image/_base.py
@@ -2,18 +2,20 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from collections.abc import Sized
+from functools import singledispatch
 from types import ModuleType
 from typing import TypeVar
 
-import dask.array as da
 import numpy as np
 import xarray as xr
 from numpy.typing import NDArray
+from sklearn.base import BaseEstimator
+from sklearn.neighbors import KNeighborsRegressor
 
 ImageType = TypeVar("ImageType", NDArray, xr.DataArray, xr.Dataset)
 
 
-class _ImagePreprocessor(ABC):
+class ImagePreprocessor(ABC):
     """
     A pre-processor for multi-channel image data in a machine learning workflow.
 
@@ -166,136 +168,22 @@ class _ImagePreprocessor(ABC):
         return flat_image
 
 
-class NDArrayPreprocessor(_ImagePreprocessor):
-    """
-    Pre-processor multi-band NumPy NDArrays.
-    """
-
-    _backend = np
-    band_dim = -1
-
-    def _flatten(self, image: NDArray) -> NDArray:
-        """Flatten the array from (y, x, bands) to (pixels, bands)."""
-        return image.reshape(-1, self.n_bands)
-
-    def unflatten(self, flat_image: NDArray, *, apply_mask=True) -> NDArray:
-        if apply_mask:
-            flat_image = self._fill_nodata(flat_image, np.nan)
-
-        return flat_image.reshape(*self.image.shape[:2], -1)
+@singledispatch
+def predict(
+    X_image: NDArray | xr.DataArray | xr.Dataset,
+    *,
+    estimator: BaseEstimator,
+    nodata_vals=None,
+) -> None:
+    raise NotImplementedError
 
 
-class DataArrayPreprocessor(_ImagePreprocessor):
-    """
-    Pre-processor for multi-band xr.DataArrays.
-    """
-
-    _backend = da
-    band_dim = 0
-
-    def _validate_nodata_vals(
-        self, nodata_vals: float | tuple[float] | NDArray | None
-    ) -> NDArray | None:
-        """
-        Get an array of NoData values in the shape (bands,) based on user input and
-        DataArray metadata.
-        """
-        # Defer to user-provided nodata values over stored attributes
-        if nodata_vals is not None:
-            return super()._validate_nodata_vals(nodata_vals)
-
-        # If present, broadcast the _FillValue attribute to all bands
-        fill_val = self.image.attrs.get("_FillValue")
-        if fill_val is not None:
-            return np.full((self.n_bands,), fill_val)
-
-        return None
-
-    def _flatten(self, image: xr.DataArray) -> xr.DataArray:
-        """Flatten the dataarray from (bands, y, x) to (pixels, bands)."""
-        # Dask can't reshape multiple dimensions at once, so transpose to swap axes
-        return image.data.reshape(self.n_bands, -1).T
-
-    def unflatten(
-        self,
-        flat_image: xr.DataArray,
-        *,
-        apply_mask=True,
-        var_names=None,
-    ) -> xr.DataArray:
-        if apply_mask:
-            flat_image = self._fill_nodata(flat_image, np.nan)
-
-        n_outputs = flat_image.shape[self.flat_band_dim]
-        # Default the variable coordinate to sequential numbers if not provided
-        var_names = var_names if var_names is not None else range(n_outputs)
-
-        # Replace the original variable coordinates and dimensions
-        dims = {**self.image.sizes, "variable": n_outputs}
-        coords = {**self.image.coords, "variable": var_names}
-        shape = list(dims.values())
-
-        return xr.DataArray(
-            # Transpose the flat image from (pixels, bands) to (bands, pixels) prior
-            # to reshaping to match the expected output.
-            flat_image.T.reshape(shape),
-            coords=coords,
-            dims=dims,
-        )
-
-
-class DatasetPreprocessor(DataArrayPreprocessor):
-    """
-    Pre-processor for multi-band xr.Datasets.
-
-    Unlike a DataArray, a Dataset will retrieve variable names and nodata values from
-    metadata, if possible.
-    """
-
-    def __init__(
-        self,
-        image: ImageType,
-        nodata_vals: float | tuple[float] | NDArray | None = None,
-        nan_fill: float | None = 0.0,
-    ):
-        # The image itself will be stored as a DataArray, but keep the Dataset for
-        # metadata like _FillValues.
-        self.dataset = image
-        super().__init__(image.to_dataarray(), nodata_vals, nan_fill)
-
-    def _validate_nodata_vals(
-        self, nodata_vals: float | tuple[float] | NDArray | None
-    ) -> NDArray | None:
-        """
-        Get an array of NoData values in the shape (bands,) based on user input and
-        Dataset metadata.
-        """
-        fill_vals = [
-            self.dataset[var].attrs.get("_FillValue") for var in self.dataset.data_vars
-        ]
-
-        # Defer to provided nodata vals first. Next, try using per-variable fill values.
-        # If at least one variable specifies a nodata value, use them all. Variables
-        # that didn't specify a fill value will be assigned None.
-        if nodata_vals is None and not all(v is None for v in fill_vals):
-            return np.array(fill_vals)
-
-        # Fall back to the DataArray logic for handling nodata
-        return super()._validate_nodata_vals(nodata_vals)
-
-    def unflatten(
-        self,
-        flat_image: xr.DataArray,
-        *,
-        apply_mask=True,
-        var_names=None,
-    ) -> xr.Dataset:
-        return (
-            super()
-            .unflatten(
-                flat_image,
-                apply_mask=apply_mask,
-                var_names=var_names,
-            )
-            .to_dataset(dim="variable")
-        )
+@singledispatch
+def kneighbors(
+    X_image: NDArray | xr.DataArray | xr.Dataset,
+    *,
+    estimator: KNeighborsRegressor,
+    nodata_vals=None,
+    **kneighbors_kwargs,
+) -> None:
+    raise NotImplementedError
